+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Version 0.19.4rc (2019-12-23)
 - enh: `base_iterator` split out `SimpleCacheMan` and made cache calls less verbose
 - enh: convert the usage of `context[break_iterator]` to raising an exception `IsitfitCliRunnerBreakIterator` and then catching it in the `mainManager`
 - enh: promote the stderr message "Will skip ... out of ... regions" to stdout and display it exactly once per service
+- enh: split the "AWS returned AccessDenied for ... out of ... regions" into 2 parts, 1 going to stdout and the other going to stderr at `--verbose` level
 
 
 Version 0.19.3 (2019-12-21)